+++ src/cpp/include/nodegui/core/Integration/integration.h
