+++ src/cpp/lib/core/Integration/integration.cpp
@@ -1,5 +1,5 @@
 
-#include "Extras/Utils/integration.h"
+#include "core/Integration/integration.h"
 
 #include <QApplication>
 #include <QFont>
